% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatiter-class.R
\name{combinatiter}
\alias{combinatiter}
\title{Factory of cobinatrics iterators}
\usage{
combinatiter(nextFunc, prevFunc, first, last, ...)
}
\arguments{
\item{nextFunc, prevFunc}{Function that takes iterator value
(and possibly other arguments) and returns the next or previous value}

\item{first, last}{First and last value of the iterator}
}
\value{
\code{combinatiter} object
}
\description{
Factory of cobinatrics iterators
}
\keyword{internal}

