% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3generic.R
\name{getFirst}
\alias{getFirst}
\title{First Value of Iterator}
\usage{
getFirst(obj, ...)
}
\arguments{
\item{obj}{an R object}

\item{...}{additional arguments}
}
\value{
iterator value, format dependes on the objects
}
\description{
\code{getFirst} is a generic function that returns the
first value of iterators
}
