% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3generic.R
\name{getLast}
\alias{getLast}
\title{Last Value of Iterator}
\usage{
getLast(obj, ...)
}
\arguments{
\item{obj}{an R object}

\item{...}{additional arguments}
}
\value{
iterator value, format dependes on the objects
}
\description{
\code{getFirst} is a generic function that returns the
last value of iterators
}
