% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3generic.R
\name{hasPrev}
\alias{hasPrev}
\title{Does This Iterator Have A Previous Element}
\usage{
hasPrev(obj, ...)
}
\arguments{
\item{obj}{an R object}

\item{...}{additional arguments}
}
\value{
Logical value indicating whether the iterator has a previous element.
}
\description{
\code{hasPrev} is a generic function that indicates if the
iterator has another element backward.
}
