% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incremental-iter.R
\name{incrementaliter}
\alias{incrementaliter}
\title{Factory of Iterators with Incremental Operations}
\usage{
incrementaliter(nextFunc, prevFunc, firstFunc, lastFunc, ...)
}
\arguments{
\item{nextFunc, prevFunc}{Functions that take iterator value
(and possibly other arguments) and returns the next or previous value}

\item{firstFunc, lastFunc}{Functions that take parameters given by `...`
and returns the first or last values of the iteration}

\item{...}{parameters that characterize the iterator}
}
\value{
\code{combinatiter} object
}
\description{
Factory of Iterators with Incremental Operations
}
