% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3generic.R
\name{prevElem}
\alias{prevElem}
\title{Get Previous Element of Iterator}
\usage{
prevElem(obj, ...)
}
\arguments{
\item{obj}{an R object}

\item{...}{additional arguments}
}
\value{
iterator value
}
\description{
\code{prevElem} is a generic funcion to move an
iterator object one step backward.
}
