% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icartes.R
\name{icartes}
\alias{icartes}
\title{Cartesean Product Iterator}
\usage{
icartes(nvec)
}
\arguments{
\item{nvec}{integer vector of number of items}
}
\value{
iterator object
}
\description{
Cartesean Product Iterator
}
\examples{
x <- icartes(c(3, 2, 4))
ct <- 0
while (hasNext(x))
{
  ct <- ct + 1
  i <- nextElem(x)
  cat(sprintf("\%3d : \%s\\n", ct, paste0(i, collapse = " ")))
}
}

