% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterators.R
\name{icomb}
\alias{icomb}
\title{Combination Iterator}
\usage{
icomb(n, k)
}
\arguments{
\item{n}{positive integer}

\item{k}{positive integer no greater than n}
}
\value{
iterator object
}
\description{
Create an iterator for all combinations k integers out of 1 through n.
}
\examples{
x <- icomb(5, 3)
ct <- 0
while (hasNext(x))
{
  ct <- ct + 1
  i <- nextElem(x)
  cat(sprintf("\%3d : \%s\\n", ct, paste0(i, collapse = " ")))
}
}
