% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterators.R
\name{iperm}
\alias{iperm}
\title{Permutation Iterator}
\usage{
iperm(n)
}
\arguments{
\item{n}{positive integer}
}
\value{
iterator object
}
\description{
Create an iterator for all permutations of integers 1 through n.
}

