% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iperm.R
\name{iperm}
\alias{iperm}
\title{Permutation Iterator}
\usage{
iperm(n, k = n)
}
\arguments{
\item{n}{positive integer}

\item{k}{positive integer}
}
\value{
iterator object
}
\description{
Create an iterator for all permutations of integers 1 through n.
}
\examples{
x <- iperm(3)
ct <- 0
while (hasNext(x))
{
  ct <- ct + 1
  i <- nextElem(x)
  cat(sprintf("\%3d : \%s\\n", ct, paste0(i, collapse = " ")))
}
}
