% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterators.R
\name{isubset}
\alias{isubset}
\title{Subset Iterator}
\usage{
isubset(n)
}
\arguments{
\item{n}{positive integer}
}
\value{
iterator object
}
\description{
Create an iterator for all subsets of integers 1 through n.
}
\examples{
x <- isubset(3)
ct <- 0
while (hasNext(x))
{
  ct <- ct + 1
  i <- nextElem(x)
  cat(sprintf("\%3d : \%s\\n", ct, paste0(i, collapse = " ")))
}
}

